backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pymongo
passlib[bcrypt]
python-jose[cryptography]
